refactor(direct-control): extract autoplay and turn types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only autoplay status/options/action result type ownership into packages/civ7-direct-control/src/play/autoplay.ts and turn-completion status/action result type ownership into packages/civ7-direct-control/src/play/turn-completion.ts. Keeps facade type re-exports in index.ts.

Preserves public type contracts and does not change autoplay command source, approval behavior, stop-settling/pause behavior, turn-completion command strings, stale notification fallback classification, runtime proof status, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance. Tactical, operation, ready, public procedure schema, and telemetry type ownership remain pending separate owner slices.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- autoplay-and-turn.test.ts public-api.test.ts
- rg -n "game play end-turn|autoplay|turn-complete|turn completion" packages/cli/test/commands packages/cli/src/commands
- bun run --cwd packages/cli test -- game.play.end-turn.test.ts game.control.test.ts game.restart.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle rg scans

Local package/type relocation proof only; no runtime/live-game, autoplay behavior, turn-completion behavior, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
