+++ openspec/changes/civ7-support-direct-control-modularization/workstream/workstream-record.md
@@ -689,6 +689,20 @@ All future agent waves must be framed before delegation:
   and telemetry type ownership remain pending separate owner slices. This is
   local package/type relocation proof only, not runtime proof, AI/hotseat
   product-path support, or procedure-core readiness.
+- Direct-control autoplay/turn types slice: completed as a narrow public type
+  ownership relocation. It moves only autoplay status/options/action result
+  types into `src/play/autoplay.ts` and turn-completion status/action result
+  types into `src/play/turn-completion.ts` while keeping facade type re-exports
+  in `index.ts` and replacing moved autoplay/turn type imports with concrete
+  owner imports. This preserves public type contracts and does not change
+  autoplay command source, approval behavior, stop-settling/pause behavior,
+  turn-completion command strings, stale notification fallback classification,
+  runtime proof status, telemetry, AI ingestion, hotseat runtime proof, CLI
+  semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4
+  matrix-row acceptance. Tactical, operation, ready, public procedure schema,
+  and telemetry type ownership remain pending separate owner slices. This is
+  local package/type relocation proof only, not runtime proof, AI/hotseat
+  product-path support, or procedure-core readiness.
 - Direct-control turn-completion slice: completed as a narrow turn-completion
   source/wrapper relocation. It moves only `getCiv7TurnCompletionStatus`,
   `sendCiv7TurnComplete`, `sendCiv7TurnUnready`, the status command builder,
